# DOC-20260426-WA0002~1/lib/api-zod/src/generated/types/createReviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jojo Collections perfume e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewBody {
  customerName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment: string;
}