# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jojo Collections perfume e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL for an object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

  # ---------- Products (public) ----------
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List approved reviews for a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Submit a review for a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }

  # ---------- Orders (public) ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    RequestUploadUrlBody:
      type: object
      properties:
        name: { type: string }
        size: { type: number }
        contentType: { type: string }
      required: [name, size, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        brand: { type: string }
        description: { type: string }
        category: { type: string, description: "e.g. Eau de Parfum, Eau de Toilette, Body Mist" }
        price: { type: number, description: "Price in dollars" }
        sizeMl: { type: integer, nullable: true }
        stock: { type: integer }
        featured: { type: boolean }
        imageUrl: { type: string, nullable: true }
        notes:
          type: array
          items: { type: string }
          description: "Fragrance notes (top, heart, base mixed)"
        topNotes: { type: string, nullable: true }
        heartNotes: { type: string, nullable: true }
        baseNotes: { type: string, nullable: true }
        averageRating: { type: number, nullable: true }
        reviewCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, name, brand, description, category, price, stock, featured, notes, reviewCount, createdAt]
...
    Review:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string, nullable: true }
        customerName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        status:
          type: string
          enum: [pending, approved, hidden]
        createdAt: { type: string, format: date-time }
      required: [id, productId, customerName, rating, comment, status, createdAt]
    CreateReviewBody:
      type: object
      properties:
        customerName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
      required: [customerName, rating, comment]
...